refactor: streamline X extraction logic and enhance floating button cleanup
- Simplified X extraction logic by moving core functions to dedicated submodules.
- Improved floating button cleanup process by consolidating event listener removals and ensuring full cleanup via the controller.
- Updated fullscreen event handling for better maintainability.
